+++ SettlerCraft-StructureAPI/src/main/java/com/chingo247/settlercraft/structureapi/platforms/bukkit/selection/HologramSelectionManager.java
@@ -20,7 +20,8 @@ import com.chingo247.settlercraft.core.Direction;
 import com.chingo247.settlercraft.core.SettlerCraft;
 import com.chingo247.settlercraft.structureapi.platforms.services.holograms.Hologram;
 import com.chingo247.settlercraft.structureapi.platforms.services.holograms.HologramsProvider;
-import com.chingo247.settlercraft.structureapi.selection.ISelectionManager;
+import com.chingo247.settlercraft.structureapi.selection.ASelectionManager;
+import com.chingo247.settlercraft.structureapi.selection.Selection;
 import com.chingo247.settlercraft.structureapi.util.WorldUtil;
 import com.chingo247.xplatform.core.APlatform;
 import com.chingo247.xplatform.core.IColors;
@@ -31,13 +32,14 @@ import com.sk89q.worldedit.entity.Player;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
+import java.util.Objects;
 import java.util.UUID;
 
 /**
  *
  * @author Chingo
  */
-public class HologramSelectionManager implements ISelectionManager{
+public class HologramSelectionManager extends ASelectionManager {
     
     private HologramsProvider hologramsProvider;
     private static HologramSelectionManager instance;
@@ -71,43 +73,30 @@ public class HologramSelectionManager implements ISelectionManager{
     @Override
     public void select(Player player, Vector pos1, Vector pos2) {
         Direction direction = WorldUtil.getDirection(player.getLocation().getYaw());
-        
         Vector pos = WorldUtil.translateLocation(pos1, direction, pos2.getBlockX(), pos2.getBlockY(), pos2.getBlockZ());
-        System.out.println("Pos: " + pos);
-        
-        selections.put(player.getUniqueId(), new HoloSelection(player, pos1, pos2, direction, pos1.getBlockX() < 0));
+        HoloSelection selection = new HoloSelection(player, pos1, pos2, direction, pos.getBlockX() < 0);
+        putSelection(selection);
     }
 
     @Override
     public void deselect(Player player) {
         UUID uuid = player.getUniqueId();
-        if (selections.get(uuid) != null) {
-            selections.get(uuid).clear();
-            selections.remove(uuid);
+        System.out.println("Has selection?");
+        if(hasSelection(player)) {
+            System.out.println("Has selection!");
+            HoloSelection selection = (HoloSelection) getSelection(uuid);
+            for(Hologram h : selection.holos) {
+                System.out.println("Delete holo!");
+                h.delete();
+            }
+            removeSelection(selection);
         }
     }
 
-    @Override
-    public boolean hasSelection(Player player) {
-        return selections.get(player.getUniqueId()) != null;
-    }
+   
 
-    @Override
-    public boolean matchesCurrentSelection(Player player, Vector pos1, Vector pos2) {
-        HoloSelection selection = selections.get(player.getUniqueId());
-        if (selection != null) {
+    private class HoloSelection extends Selection {
 
-            // is it the same structure and at the same position?
-            return  selection.pos1.equals(pos1)
-                    && selection.pos2.equals(pos2);
-        }
-        return false;
-    }
-
-    private class HoloSelection {
-
-        private final Vector pos1;
-        private final Vector pos2;
 
         private final Hologram[] holos;
 
@@ -115,33 +104,31 @@ public class HologramSelectionManager implements ISelectionManager{
         private static final int X_AXIS = 1;
         private static final int Y_AXIS = 2;
         private static final int Z_AXIS = 3;
-        private static final int END = 4;
 
         public HoloSelection(final Player whoCanSee, final Vector pos1, final Vector pos2, Direction direction, boolean reverse) {
-            this.pos1 = pos1;
-            this.pos2 = pos2;
-            this.holos = new Hologram[5];
+            super(whoCanSee.getUniqueId(), pos1, pos2);
+            this.holos = new Hologram[4];
 
             final Vector self;
             switch (direction) {
                 case WEST:
-                    self = WorldUtil.translateLocation(pos1, direction, -0.5, 1, -0.5);
+                    self = WorldUtil.translateLocation(pos1, direction, 0, 1.2, -0.5);
                     break;
                 case SOUTH:
-                    self = WorldUtil.translateLocation(pos1, direction, -0.5, 1, 0.5);
+                    self = WorldUtil.translateLocation(pos1, direction, -0.5, 1.2, 0.5);
                     break;
                 case NORTH:
-                    self = WorldUtil.translateLocation(pos1, direction, 0.5, 1, -0.5);
+                    self = WorldUtil.translateLocation(pos1, direction, 0.5, 1.2, -0.5);
                     break;
                 case EAST:
-                    self = WorldUtil.translateLocation(pos1, direction, 0.5, 1, 0.5);
+                    self = WorldUtil.translateLocation(pos1, direction, 0.5,  1.2, 0.5);
                     break;
                 default:
                     return;
             }
 
             final Vector xLoc;
-            if (!reverse) {
+            if (reverse) {
                 xLoc = WorldUtil.translateLocation(self, direction, 2, 1, 0);
             } else {
                 xLoc = WorldUtil.translateLocation(self, direction, -2, 1, 0);
@@ -162,7 +149,7 @@ public class HologramSelectionManager implements ISelectionManager{
                     Hologram boxHolo = hologramsProvider.createHologram(
                             plugin.getName(), 
                             whoCanSee.getWorld(), 
-                            new Vector(self.getBlockX(), self.getBlockY(), self.getBlockZ())
+                            new Vector(self.getX(), self.getY(), self.getZ())
                     );
                     boxHolo.addLine(colors.green() + "[X]");
                     holos[SELF] = boxHolo;
@@ -171,27 +158,27 @@ public class HologramSelectionManager implements ISelectionManager{
                     Hologram xAxisHolo = hologramsProvider.createHologram(
                             plugin.getName(), 
                             whoCanSee.getWorld(), 
-                            new Vector(xLoc.getBlockX(), xLoc.getBlockY(), xLoc.getBlockZ())
+                            new Vector(xLoc.getX(), xLoc.getY(), xLoc.getZ())
                     );
-                    boxHolo.addLine(colors.yellow() + "X-AXIS" + colors.reset() + "+" + Math.abs(pos2.subtract(pos1).getBlockX()));
+                    xAxisHolo.addLine(colors.yellow() + "X-AXIS" + colors.reset() + "+" + Math.abs(pos2.subtract(pos1).getX()));
                     holos[X_AXIS] = xAxisHolo;
 //                           
                     // Y-AXIS Holo
                     Hologram yAxisHolo = hologramsProvider.createHologram(
                             plugin.getName(), 
                             whoCanSee.getWorld(), 
-                            new Vector(yLoc.getBlockX(), yLoc.getBlockY(), yLoc.getBlockZ())
+                            new Vector(yLoc.getX(), yLoc.getY(), yLoc.getZ())
                     );
-                    boxHolo.addLine(colors.yellow() + "Y-AXIS" + colors.reset() + "+" + Math.abs(pos2.subtract(pos1).getBlockY()));
+                    yAxisHolo.addLine(colors.yellow() + "Y-AXIS" + colors.reset() + "+" + Math.abs(pos2.subtract(pos1).getY()));
                     holos[Y_AXIS] = yAxisHolo;
                     
                     // Z-AXIS Holo
                     Hologram zAxisHolo = hologramsProvider.createHologram(
                             plugin.getName(), 
                             whoCanSee.getWorld(), 
-                            new Vector(zLoc.getBlockX(), zLoc.getBlockY(), zLoc.getBlockZ())
+                            new Vector(zLoc.getX(), zLoc.getY(), zLoc.getZ())
                     );
-                    boxHolo.addLine(colors.yellow() + "Z-AXIS" + colors.reset() + "+" + Math.abs(pos2.subtract(pos1).getBlockZ()));
+                    zAxisHolo.addLine(colors.yellow() + "Z-AXIS" + colors.reset() + "+" + Math.abs(pos2.subtract(pos1).getZ()));
                     holos[Z_AXIS] = zAxisHolo;
                 }
             });
@@ -199,6 +186,34 @@ public class HologramSelectionManager implements ISelectionManager{
             
 
         }
+
+             @Override
+    public int hashCode() {
+        int hash = 5;
+        hash = 47 * hash + Objects.hashCode(this.start);
+        hash = 47 * hash + Objects.hashCode(this.end);
+        return hash;
+    }
+
+    @Override
+    public boolean equals(Object obj) {
+        if (obj == null) {
+            return false;
+        }
+        if (getClass() != obj.getClass()) {
+            return false;
+        }
+        final Selection other = (Selection) obj;
+        if (!Objects.equals(this.start, other.getStart())) {
+            return false;
+        }
+        if (!Objects.equals(this.end, other.getEnd())) {
+            return false;
+        }
+        return true;
+    }
+        
+        
 
         private void clear() {
             final APlatform platform = SettlerCraft.getInstance().getPlatform();
@@ -210,7 +225,9 @@ public class HologramSelectionManager implements ISelectionManager{
                 @Override
                 public void run() {
                     for (Hologram holo : holos) {
-                        holo.delete();
+                        if(holo != null) {
+                            holo.delete();
+                        }
                     }
                 }
             });
@@ -220,7 +237,9 @@ public class HologramSelectionManager implements ISelectionManager{
     
     public void clearAll() {
         for(HoloSelection s : selections.values()) {
-            s.clear();
+            for(Hologram h : s.holos) {
+                h.delete();
+            }
         }
     }
